requirements.txt
pandas~=2.2.3
PyYAML~=6.0.2
flet~=0.27.1
openpyxl~=3.1.2